# Speedup pow_of_ten/1