--- a/rebar.config
+++ b/rebar.config
@@ -1,5 +1,5 @@
 {xref_checks, [undefined_function_calls]}.
-
+{project_plugins, [rebar3_bench]}.
 {plugins, [rebar3_proper]}.
 {profiles,
     [{test, [